+++ R/client-connectCloud.R
@@ -246,16 +246,20 @@ connectCloudClient <- function(service, authInfo) {
         primary_file = primaryFile
       )
 
-      secrets <- unname(Map(
-        function(name, value) {
-          list(
-            name = name,
-            value = value
-          )
-        },
-        envVars,
-        Sys.getenv(envVars)
-      ))
+      secrets <- if (length(envVars) > 0) {
+        unname(Map(
+          function(name, value) {
+            list(
+              name = name,
+              value = value
+            )
+          },
+          envVars,
+          Sys.getenv(envVars)
+        ))
+      } else {
+        list()
+      }
 
       json <- list(
         account_id = accountId,
@@ -287,16 +291,20 @@ connectCloudClient <- function(service, authInfo) {
         path <- paste0(path, "?new_bundle=true")
       }
 
-      secrets <- unname(Map(
-        function(name, value) {
-          list(
-            name = name,
-            value = value
-          )
-        },
-        envVars,
-        Sys.getenv(envVars)
-      ))
+      secrets <- if (length(envVars) > 0) {
+        unname(Map(
+          function(name, value) {
+            list(
+              name = name,
+              value = value
+            )
+          },
+          envVars,
+          Sys.getenv(envVars)
+        ))
+      } else {
+        list()
+      }
 
       json <- list(
         secrets = secrets,